+++ src/lib.rs
@@ -571,7 +571,7 @@ pub mod prelude {
     pub use crate::progress::ProgressPrinter;
     pub use crate::structures::atom::Atom;
     pub use crate::structures::dimension::Dimension;
-    pub use crate::structures::element::{Element, SupportedElements};
+    pub use crate::structures::element::SupportedElements;
     pub use crate::structures::shape::{Cylinder, Rectangular, Shape, Sphere};
     pub use crate::structures::simbox::SimBox;
     pub use crate::structures::vector3d::Vector3D;